svelte-app/pyproject.toml
[tool.poetry]
name = "hebrew"
version = "0.1.0"
description = ""
authors = ["sethbam9 <54952081+sethbam9@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
whitenoise = "^6.4.0"
djangorestframework = "^3.14.0"
pytz = "^2023.3"
mypy = "^1.2.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
